requirements.txt
# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0
einops>=0.7.0
transformers>=4.35.0

# Mamba (State Space Model)
mamba-ssm>=1.2.0
causal-conv1d>=1.1.0

# Vision
timm>=0.9.0
opencv-python>=4.8.0
pillow>=10.0.0
imageio>=2.31.0
albumentations>=1.3.0

# Training
accelerate>=0.24.0
wandb>=0.15.0
pytorch-lightning>=2.0.0
webdataset>=0.2.0

# Optimization
onnx>=1.14.0
onnxruntime-gpu>=1.16.0
tensorrt>=8.6.0
torch-tensorrt>=1.4.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
h5py>=3.9.0
zarr>=2.16.0

# API & Deployment
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0
pydantic>=2.0.0

# Frontend Bridge
fastapi-socketio>=0.0.10
python-multipart>=0.0.6

# Utilities
tqdm>=4.65.0
omegaconf>=2.3.0
hydra-core>=1.3.0
rich>=13.4.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Quantization
bitsandbytes>=0.41.0
quanto>=0.0.10